requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg[binary]==3.2.2
psycopg-pool==3.2.2

# AI (OpenAI SDK with Groq backend - OpenAI-compatible API)
openai>=1.0.0

# HTTP Client
httpx==0.26.0

# Utils
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1

# Timezone
pytz==2024.1
tzlocal==5.2




